+++ src/com/towson/wavyleaf/Main.java
@@ -5,6 +5,7 @@ import android.app.AlertDialog;
 import android.app.Dialog;
 import android.app.NotificationManager;
 import android.app.PendingIntent;
+import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
@@ -15,6 +16,9 @@ import android.database.sqlite.SQLiteDatabase;
 import android.graphics.Color;
 import android.graphics.Typeface;
 import android.graphics.drawable.Drawable;
+import android.location.Location;
+import android.location.LocationListener;
+import android.location.LocationManager;
 import android.os.Build;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
@@ -42,6 +46,7 @@ public class Main extends SherlockActivity implements OnClickListener {
 	NotificationManager nm;
 //	public CountDownTimer ctd;
 	public AlarmManager am;
+	public LocationApplication locationData;
 	
 	@Override
 	public void onCreate(Bundle bundle) {
@@ -54,7 +59,9 @@ public class Main extends SherlockActivity implements OnClickListener {
 		initLayout();
 		determineButtonDrawable();
 		toggleFirstRun();
+		findUsersLocation();
     }
+	
 	
 	protected void initLayout() {
 		Typeface tf_light = Typeface.createFromAsset(getAssets(), "fonts/roboto_light.ttf");
@@ -419,6 +426,11 @@ public class Main extends SherlockActivity implements OnClickListener {
 		Intent helpIntent = new Intent(this, Help.class);
 		helpIntent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP | Intent.FLAG_ACTIVITY_NEW_TASK);
 		startActivity(helpIntent);
+	}
+	
+	private void findUsersLocation(){
+		locationData = (LocationApplication) getApplication();
+		locationData.init();
 	}
 		
 }